refactor(uffd-tests): extract rpc wire types/client/barrier registry to internal/rpcharness
Move the parts of the test harness that don't touch *Userfaultfd
internals into a new internal subpackage:

  packages/.../uffd/userfaultfd/internal/rpcharness/
      wire.go     — Bootstrap/PageStates/FaultBarrier/Token args+replies
      client.go   — typed Client wrapping *rpc.Client (Bootstrap, Pause,
                    Resume, PageStates, InstallBarrier, WaitFaultHeld,
                    ReleaseFault, Shutdown, Close)
      barriers.go — Registry + Point + HookFor

The RPC service implementations stay as _test.go in userfaultfd
(they need access to unexported pageStateEntries, settleRequests,
pageTracker, defaultCopyMode etc.). They become smaller — wire types
no longer locally defined; barrier registry consumed via interface
boundary; testHandler holds a single *rpcharness.Client instead of
{*rpc.Client, servePause, serveResume, pageStates} func fields.

internal/ scoping ensures the testutil package can never leak into a
production import path.

Test-only.
